Fix dinding moving
