+++ airflow/contrib/example_dags/example_dingding_operator.py
@@ -22,7 +22,7 @@ This is an example dag for using the DingdingOperator.
 from datetime import timedelta
 
 from airflow import DAG
-from airflow.providers.dindding.operators.dingding import DingdingOperator
+from airflow.providers.dingding.operators.dingding import DingdingOperator
 from airflow.utils.dates import days_ago
 
 args = {
