+++ docs/howto/operator/dingding.rst
@@ -34,7 +34,7 @@ To use this operators, you must do a few things:
 Basic Usage
 ^^^^^^^^^^^
 
-Use the :class:`~airflow/contrib/operators/dingding_operator.DingdingOperator`
+Use the :class:`~airflow.contrib.operators.dingding_operator.DingdingOperator`
 to send Dingding message:
 
 .. exampleinclude:: ../../../airflow/contrib/example_dags/example_dingding_operator.py
